+++ src/main/java/ru/kurbanmagomedov/springCRUD/service/UserService.java
@@ -9,11 +9,11 @@ public interface UserService {
 
     void saveUser(User user);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     void setUser(User user);
 
-    void removeUser(int id);
+    void removeUser(Long id);
 
     List<User> getAllUsers();
 }
